pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-semantic-archive"
version = "0.1.0"
description = "Hybrid visual semantic archive with CLIP, faces and FTS."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "VSA Maintainers" }]
dependencies = [
  "streamlit>=1.32,<2.0",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "httpx>=0.27,<1.0",
  "chromadb>=0.4.24,<0.5",
  "open_clip_torch>=2.24,<3.0",
  "insightface>=0.7.3,<0.8",
  "onnxruntime-gpu>=1.17,<1.21",
  "torch>=2.2,<2.4",
  "opencv-python>=4.8,<5.0",
  "numpy>=1.26,<2.0",
  "Pillow>=10.2,<12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "pytest-cov>=4.1,<6.0",
  "ruff>=0.4,<0.7",
  "black>=24.3,<25.0",
  "mypy>=1.8,<2.0",
  "types-requests",
]

[project.scripts]
vsa = "core.cli:main"

[tool.setuptools]
packages = ["core"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "PL"]
ignore = [
  "PLR0913",
  "PLR0912",
  "PLR0915",
  "PLR2004",
]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"
